Crate parallel_primitives

Modules

network

Structs

PriceWithDecimal

Enums

CurrencyId

DataProviderId

Constants

SECONDS_PER_YEAR

Traits

EmergencyPriceFeeder

ExchangeRateProvider

PriceFeeder

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex
The type for looking up accounts. We don’t expect more than 4 billion of them.

AccountPublic
Alias to the public key used for this chain, actually a MultiSigner. Like the signature, this also isn’t a fixed size when encoded, as different cryptos have different size public keys.

Amount
Signed version of Balance

AuraId
An Aura authority identifier using S/R 25519 as its crypto.

Balance
Balance of an account.

BlockNumber
An index to a block.

EraIndex
Index of era in relaychain.

Hash
A hash of some data used by the chain.

Index
Index of a transaction in the chain. 32-bit should be plenty.

Liquidity

Moment
An instant or duration in time.

Price

PriceDetail

Rate
The fixed point number

Ratio
The fixed point number, range from 0 to 1.

Shortfall

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.

TimeStampedPrice

Timestamp