Module parallel_primitives::ump

Structs

BalancesTransferAllCall

Relaychain balances.transfer_all call arguments

BalancesTransferKeepAliveCall

Relaychain balances.transfer_keep_alive call arguments

CrowdloansAddMemoCall

CrowdloansContributeCall

CrowdloansWithdrawCall

ProxyAddProxyCall

ProxyProxyCall

ProxyRemoveProxyCall

StakingBondCall

Relaychain staking.bond call arguments

StakingBondExtraCall

Relaychain staking.bond_extra call arguments

StakingNominateCall

Relaychain staking.nominate call arguments

StakingPayoutStakersCall

Relaychain staking.payout_stakers call arguments

StakingRebondCall

Relaychain staking.rebond call arguments

StakingUnbondCall

Relaychain staking.unbond call arguments

StakingWithdrawUnbondedCall

Relaychain staking.withdraw_unbonded call arguments

SystemRemarkCall

UtilityAsDerivativeCall

Relaychain utility.as_derivative call arguments

UtilityBatchAllCall

Relaychain utility.batch_all call arguments

XcmWeightMisc

The xcm weight when execute ump call wrapped in xcm message

Enums

BalancesCall

CrowdloansCall

KusamaCall

PolkadotCall

ProxyCall

ProxyType

The type used to represent the kinds of proxying allowed.

RewardDestination

A destination account for payment.

StakingCall

SystemCall

UtilityCall