Crate parallel_primitives

Modules

network

tokens

Structs

ParaId

Unique identifier of a parachain.

PersistedValidationData

The validation data provides information about how to create the inputs for validation of a candidate. This information is derived from the chain state and will vary from para to para, although some fields may be the same for every para.

Enums

ArithmeticKind

DataProviderId

Constants

SECONDS_PER_YEAR

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them.

AccountPublic

Alias to the public key used for this chain, actually a MultiSigner. Like the signature, this also isn’t a fixed size when encoded, as different cryptos have different size public keys.

Amount

Signed version of Balance

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Balance

Balance of an account.

BlockNumber

An index to a block.

BridgeInterval

ChainId

ChainNonce

CurrencyId

DerivativeIndex

EraIndex

Hash

A hash of some data used by the chain.

Index

Index of a transaction in the chain. 32-bit should be plenty.

LeasePeriod

Liquidity

Moment

An instant or duration in time.

Price

PriceDetail

Rate

The fixed point number

Ratio

The fixed point number, range from 0 to 1.

Shortfall

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

StreamId

TimeStampedPrice

Timestamp

TrieIndex

VaultId